backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
openai==1.3.0
anthropic==0.7.0
pyyaml==6.0.1
lxml==4.9.3
python-dotenv==1.0.0
httpx==0.25.2
asyncio==3.4.3
typing-extensions==4.8.0

# For s(CASP) integration (will need custom installation)
# swiplserver==0.1.12

# For testing
pytest==7.4.3
pytest-asyncio==0.21.1